+++ packages/dataviews/src/filter-and-sort-data-view.ts
@@ -375,16 +375,32 @@ export function filterSortAndPaginate< Item >(
 	}
 
 	// Handle sorting.
-	if ( view.sort ) {
-		const fieldId = view.sort.field;
-		const fieldToSort = _fields.find( ( field ) => {
-			return field.id === fieldId;
-		} );
-		if ( fieldToSort ) {
-			filteredData.sort( ( a, b ) => {
-				return fieldToSort.sort( a, b, view.sort?.direction ?? 'desc' );
+	if ( view.sort || view.groupByField ) {
+		filteredData.sort( ( a, b ) => {
+			// Sort by the group by field.
+			const groupByField = _fields.find( ( field ) => {
+				return field.id === view.groupByField;
 			} );
-		}
+			if ( view.groupByField && groupByField ) {
+				const groupCompare = groupByField.sort( a, b, 'asc' );
+
+				// If items are in different groups, return the group comparison result.
+				// Otherwise, fall back to sorting by the sort field.
+				if ( groupCompare !== 0 ) {
+					return groupCompare;
+				}
+			}
+
+			// Sort by the sort field.
+			const sortByField = _fields.find( ( field ) => {
+				return field.id === view?.sort?.field;
+			} );
+			if ( view.sort && sortByField ) {
+				return sortByField.sort( a, b, view.sort?.direction ?? 'desc' );
+			}
+
+			return 0;
+		} );
 	}
 
 	// Handle pagination.