+++ packages/dataviews/CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Features
 
+- Add support for grouping items by a field in the `grid` layout by introducing a `groupByField` prop in the View object.
 - Add support for free composition in the `DataViews` component by exporting subcomponents: `<DataViews.ViewConfig />`, `<DataViews.Search />`, `<DataViews.Pagination />`, `<DataViews.LayoutSwitcher />`, `<DataViews.Layout />`, `<DataViews.FiltersToggle />`, `<DataViews.Filters />`, `<DataViews.BulkActionToolbar />`.
 - `select`, `text`, `email` controls: add `help` support from the field `description` prop.
 - `text`, `email` Edit control: add `help` support from the field `description` prop.